TeradataConnection-class {teradatasql}	R Documentation

TeradataConnection class

Description

TeradataConnection class

Insert rows into an existing table dbAppendTable(conn, name, value, row.names = NULL, ...)

Begin a transaction dbBegin(conn, ...)

Commit the current transaction dbCommit(conn, ...)

Determine the SQL data type of an R object dbDataType(dbObj, obj, ...)

Close the connection dbDisconnect(conn, ...)

Execute a SQL request dbExecute(conn, statement, ...)

Determine whether a table exists dbExistsTable(conn, name, ...)

Obtain metadata from the connection dbGetInfo(dbObj, ...)

Execute a query dbGetQuery(conn, statement, ...)

Determine whether the connection is usable dbIsValid(dbObj, ...)

Obtain a table's column names dbListFields(conn, name, ...)

Obtain the list of databases, tables, and views in the system dbListObjects(conn, prefix, ...)

Obtain the list of tables and views in the current database dbListTables(conn, ...)

Drop a table dbRemoveTable(conn, name, ...)

Rollback the current transaction dbRollback(conn, ...)

Prepare or execute a query dbSendQuery(conn, statement, ...)

Prepare or execute a SQL request dbSendStatement(conn, statement, ...)

Copy data frame to table dbWriteTable(conn, name, value, ...)

Compose SQL request to create a table sqlCreateTable(con, table, fields, ...)

Usage

## S4 method for signature 'TeradataConnection'
dbAppendTable(conn, name, value, ..., row.names = NULL)

## S4 method for signature 'TeradataConnection'
dbBegin(conn, ...)

## S4 method for signature 'TeradataConnection'
dbCommit(conn, ...)

## S4 method for signature 'TeradataConnection'
dbDataType(dbObj, obj, ...)

## S4 method for signature 'TeradataConnection'
dbDisconnect(conn, ...)

## S4 method for signature 'TeradataConnection,character'
dbExecute(conn, statement, params = NULL, immediate = TRUE, ...)

## S4 method for signature 'TeradataConnection,character'
dbExistsTable(conn, name, ...)

## S4 method for signature 'TeradataConnection,Id'
dbExistsTable(conn, name, ...)

## S4 method for signature 'TeradataConnection'
dbGetInfo(dbObj, ...)

## S4 method for signature 'TeradataConnection,character'
dbGetQuery(conn, statement, n = -1, params = NULL, immediate = TRUE, ...)

## S4 method for signature 'TeradataConnection'
dbIsValid(dbObj, ...)

## S4 method for signature 'TeradataConnection,character'
dbListFields(conn, name, ...)

## S4 method for signature 'TeradataConnection,Id'
dbListFields(conn, name, ...)

## S4 method for signature 'TeradataConnection'
dbListObjects(conn, prefix = NULL, ...)

## S4 method for signature 'TeradataConnection'
dbListTables(conn, ...)

## S4 method for signature 'TeradataConnection,character'
dbRemoveTable(conn, name, fail_if_missing = TRUE, ...)

## S4 method for signature 'TeradataConnection,Id'
dbRemoveTable(conn, name, fail_if_missing = TRUE, ...)

## S4 method for signature 'TeradataConnection'
dbRollback(conn, ...)

## S4 method for signature 'TeradataConnection,character'
dbSendQuery(conn, statement, params = NULL, immediate = NA, ...)

## S4 method for signature 'TeradataConnection,character'
dbSendStatement(conn, statement, params = NULL, immediate = NA, ...)

## S4 method for signature 'TeradataConnection,character'
dbWriteTable(
  conn,
  name,
  value,
  row.names = FALSE,
  overwrite = FALSE,
  append = FALSE,
  field.types = NULL,
  temporary = FALSE,
  ...
)

## S4 method for signature 'TeradataConnection,Id'
dbWriteTable(
  conn,
  name,
  value,
  row.names = FALSE,
  overwrite = FALSE,
  append = FALSE,
  field.types = NULL,
  temporary = FALSE,
  ...
)

## S4 method for signature 'TeradataConnection'
sqlCreateTable(con, table, fields, row.names = NA, temporary = FALSE, ...)

Arguments

conn
A TeradataConnection object

name
A character string specifying a table name

value
A data frame

row.names
Either TRUE, FALSE, NA or a string. If TRUE, translate row names to a column named "row_names". If FALSE, never translate row names. If NA, translate row names only if they are a character vector.

dbObj
A TeradataConnection object

obj
An R object whose SQL type we want to determine

statement
A character string containing the SQL request text

prefix
A database name or NULL.

temporary
If TRUE, compose a SQL request to create a volatile table.

con
A TeradataConnection object

table
A character string specifying a table name

fields
Either a character vector or a data frame A named character vector: Names are column names, values are types. A data frame: field types are generated using dbDataType().

Value

TRUE invisibly

TRUE invisibly

A string giving the SQL data type for the R object

TRUE invisibly

A TeradataResult object

TRUE if the specified table or view exists, FALSE otherwise

A named list

A TeradataResult object

TRUE to indicate that the connection is usable, FALSE otherwise

A character vector of the specified table's column names, in order

A data frame with columns table (type list of Id) and is_prefix (type logical). When NULL is specified for prefix, the data frame contains one row for each database, table, and view in the system. When a database name is specified for prefix, the data frame contains one row for each table and view in the specified database. Data frame rows for each database have is_prefix == TRUE. Data frame rows for each table and view have is_prefix == FALSE.

A character vector of all table names and view names in the current database

TRUE invisibly

TRUE invisibly

A TeradataResult object

A TeradataResult object

TRUE invisibly

A string containing the SQL request text

[Package teradatasql version 17.20.0.24 Index]