TeradataResult-class {teradatasql}	R Documentation

TeradataResult class

Description

TeradataResult class

Bind values to prepared statement parameter markers dbBind(res, params, ...)

Close the result dbClearResult(res, ...)

Obtain result column metadata dbColumnInfo(res, ...)

Fetch rows dbFetch(res, n = -1, ...) An attempt to fetch from a closed result set raises an error.

Obtain the number of rows fetched dbGetRowCount(res, ...)

Obtain the number of rows affected by the SQL statement dbGetRowsAffected(res, ...)

Obtain the SQL request text dbGetStatement(res, ...)

Determine whether the entire result has been fetched dbHasCompleted(res, ...)

Determine whether the result is usable dbIsValid(dbObj, ...)

Advances to the next result set. dbNextResult(res, ...)

Sets the fetched row count for testing. setFetchedRowCount(res, ...)

Usage

## S4 method for signature 'TeradataResult'
dbBind(res, params, ...)

## S4 method for signature 'TeradataResult'
dbClearResult(res, ...)

## S4 method for signature 'TeradataResult'
dbColumnInfo(res, ...)

## S4 method for signature 'TeradataResult'
dbFetch(res, n = -1, ...)

## S4 method for signature 'TeradataResult'
dbGetRowCount(res, ...)

## S4 method for signature 'TeradataResult'
dbGetRowsAffected(res, ...)

## S4 method for signature 'TeradataResult'
dbGetStatement(res, ...)

## S4 method for signature 'TeradataResult'
dbHasCompleted(res, ...)

## S4 method for signature 'TeradataResult'
dbIsValid(dbObj, ...)

## S4 method for signature 'TeradataResult'
dbNextResult(res, ...)

## S4 method for signature 'TeradataResult'
setFetchedRowCount(res, u, ...)

Arguments

res
A TeradataResult object

params
Parameter values to bind

n
Maximum number of rows to fetch. Specify n = -1 or n = Inf to fetch all remaining rows. If the n argument is not an atomic whole number greater or equal to -1 or Inf, an error is raised.

dbObj
A TeradataResult object

u
The fetched row count

Value

The TeradataResult invisibly

TRUE invisibly

A data frame with columns name (type character), Sclass (R data type), type (type character), len (type integer), precision (type integer), scale (type integer), nullOK (type logical). The data frame contains one row for each result column.

A data frame with columns and rows corresponding to result set columns and rows. If the SQL statement returned an empty result set, the data frame will have fully-typed columns but zero rows. If the SQL statement did not return a result set, the data frame will have zero columns and zero rows. The column types of the returned data frame depend on the data returned: integer for 32-bit integer values bit64::as.integer64() for 64-bit integer values numeric for numbers with a fractional component character for text raw for blobs as.Date() for dates hms::as_hms() for times as.POSIXct() for timestamps NA for SQL NULL values

The number of rows fetched so far

The number of rows affected by the SQL statement

The SQL request text

TRUE if all rows have been fetched, FALSE otherwise

TRUE to indicate that the result is usable, FALSE otherwise

TRUE if there are more result sets; otherwise FALSE

TRUE invisibly

[Package teradatasql version 17.20.0.24 Index]